msm8226-common: Increase camera open retries
Cameras can be opened only after the camera server has configured
them. Since Android N the framework usually tries to access the
cameras to determine whether they have a flashlight too early,
causing the flashlight tile not to work. Try to open the camera
a few times before giving up to prevent this.

Note that mm-qcamera-daemon already tries to open the camera twice
with 20ms of delay, so the total number of attempts is two times
the one defined here.

Change-Id: I528059101dd768835d6d2310ac690aa2f89fa9d1